Move CMCD API files
