+++ src/api/cmcd/CmcdConfiguration.ts
