+++ src/api/source/barrel.ts
@@ -1,6 +1,5 @@
 export * from './ads/barrel';
 export * from './analytics/barrel';
-export * from './cmcd/barrel';
 export * from './drm/barrel';
 export * from './dash/barrel';
 export * from './hls/barrel';
